requirements.txt
django-mysql==3.8.1
django-rest-swagger==2.2.0
djangorestframework==3.11.1
PyMySQL==0.10.1
python-dotenv==0.14.0
python-http-client==3.3.1
pytz==2020.1
sendgrid==3.6.5
sendgrid-django==4.2.0
django-cors-headers==3.5.0
requests==2.24.0
drf-yasg==1.17.1
mysqlclient==2.0.1
africastalking==1.2.4